temp commit

+++ src/main/java/work/szczepanskimichal/project13snippetapp/snippet/SnippetService.java
@@ -1,18 +1,25 @@
 package work.szczepanskimichal.project13snippetapp.snippet;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Service;
+import org.springframework.ui.Model;
 import work.szczepanskimichal.project13snippetapp.tag.Tag;
 import work.szczepanskimichal.project13snippetapp.tag.TagService;
+import work.szczepanskimichal.project13snippetapp.user.CurrentUser;
 import work.szczepanskimichal.project13snippetapp.user.User;
 import work.szczepanskimichal.project13snippetapp.user.UserService;
+import work.szczepanskimichal.project13snippetapp.utils.Languages;
+import work.szczepanskimichal.project13snippetapp.utils.UtilLists;
 
+import javax.transaction.Transactional;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
+@Transactional
 @Service
 @RequiredArgsConstructor
 public class SnippetService {
@@ -20,6 +27,22 @@ public class SnippetService {
     private final SnippetRepository snippetRepository;
     private final UserService userService;
     private final TagService tagService;
+    private final UtilLists utilLists;
+
+    public String returnUserDashboard(CurrentUser currentUser, String folder, Long id, Model model){
+        model.addAttribute("currentUser", currentUser.getUser());
+        List<String> folderList = findAllUserFolders(currentUser.getUser().getEmail());
+        model.addAttribute("folderList", folderList);
+        if (folder != null) {
+            List<Snippet> snippetList = findAllUserSnippetsByFolder(folder, currentUser.getUser());
+            model.addAttribute("snippetList", snippetList);
+            if (id != null){
+                Snippet currentSnippet = getUserSnippetByID(id);
+                model.addAttribute("currentSnippet", currentSnippet);
+            }
+        }
+        return "user/dashboard";
+    }
 
     public List<Snippet> findAllUserSnippets(String email) {
         return snippetRepository.findAllUserSnippets(email);
@@ -29,43 +52,63 @@ public class SnippetService {
         return snippetRepository.findSnippetsByFolderAndOwner(folder, owner);
     }
 
-    public List<String> findAllUserFolders(String email) { return snippetRepository.findAllFoldersOfUser(email); }
+    public List<String> findAllUserFolders(String email) {
+        return snippetRepository.findAllFoldersOfUser(email);
+    }
 
     public void saveSnippet(Snippet snippet, Long id) {
         snippet.setOwner(userService.findByUserId(id));
         snippetRepository.save(snippet);
     }
 
-    public Snippet getUserSnippetByID(Long id) { return snippetRepository.findById(id).orElse(null); }
+    public Snippet getUserSnippetByID(Long id) {
+        return snippetRepository.getById(id);
+    }
 
-    public void deleteUserSnippetById(Long id) {
-        snippetRepository.delete(getUserSnippetByID(id));
+    public String deleteUserSnippetById(Long id, CurrentUser currentUser) {
+        if (snippetRepository.existsById(id)) {
+            if (getUserSnippetByID(id).getOwner().getId() == currentUser.getUser().getId()) {
+                // add confirmation in popup or another widok
+                snippetRepository.delete(getUserSnippetByID(id));
+                return "redirect:/user/dashboard";
+            }
+        }
+        return "public/snippetErrorDeleted";
     }
 
     public void setSnippetFolder(Snippet snippet, String inputtedFolder) {
-        if(inputtedFolder != null && inputtedFolder != "") {
+        if (inputtedFolder != null && inputtedFolder != "") {
             snippet.setFolder(inputtedFolder);
         }
     }
 
     public void setSnippetTagsAndUserTags(Snippet snippet, String inputtedTags, User currentUser) {
-        if(inputtedTags != null && inputtedTags != "") {
-            List<String> newTagStrings = Arrays.asList(inputtedTags.trim().split("\\s*,+\\s*,*\\s*"));
-
-
-            List<Tag> currentSnippetTags = (snippet.getTags() != null ? snippet.getTags() : new ArrayList<>());
-            List<Tag> newTags = newTagStrings.stream()
-                    .map(s -> new Tag(null, s, null))
-                    .collect(Collectors.toList());
-
-            List<Tag> userTags = (currentUser.getTags() != null ? currentUser.getTags() : new ArrayList<>());
-            for(Tag tag : newTags) {
-                currentSnippetTags.add(tag);
-                userTags.add(tag);
-                tagService.save(tag);
-            }
-            userService.update(currentUser);
+        List<String> newTagStrings = new ArrayList<>();
+        if (inputtedTags != null && inputtedTags != "") {
+            newTagStrings = Arrays.asList(inputtedTags.trim().split("\\s*,+\\s*,*\\s*"));
         }
+        List<Tag> currentTags = (snippet.getTags() != null ? snippet.getTags() : new ArrayList<>());
+
+        List<Tag> userTags = (currentUser.getTags() != null ? currentUser.getTags() : new ArrayList<>());
+        for (String string : newTagStrings) {
+            Tag newTag = Tag.builder().tagName(string).tagColor(null).build();
+            currentTags.add(newTag);
+            userTags.add(newTag);
+            tagService.save(newTag);
+        }
+        snippet.setTags(currentTags);
+        snippetRepository.save(snippet);
+
+    }
+
+    public void getProgrammingLanguagesAndUserFoldersAndUserTags(@AuthenticationPrincipal CurrentUser currentUser, Model model) {
+        List<String> programmingLanguages = Languages.getLanguages();
+        List<String> folderList = (findAllUserFolders(currentUser.getUser().getEmail()).isEmpty() || findAllUserFolders(currentUser.getUser().getEmail()) == null)
+                ? utilLists.getDefaultFolder() : findAllUserFolders(currentUser.getUser().getEmail());
+        List<Tag> userTags = currentUser.getUser().getTags();
+        model.addAttribute("programmingLanguages", programmingLanguages);
+        model.addAttribute("folderList", folderList);
+        model.addAttribute("userTags", userTags);
     }
 
     //TODO method for adding all values into snippet instead of controller method .... balagan
